+++ app/admin/arbre_helpers.rb
@@ -1,16 +1,4 @@
 module ArbreHelpers
-  def self.fields_for_replaces(context, form, assoc)
-    Appsignal.instrument("render_fields_for_replaces") do
-      context.instance_eval do
-        if replaceable = context.resource.person.send(assoc).current.presence
-          form.input :replaces, collection: replaceable
-          form.input :copy_attachments,
-            label: "Move attachments of replaced #{assoc} to the new one"
-        end
-      end
-    end
-  end
-
   def self.panel_grid(context, objects, &block)
     Appsignal.instrument("render_#{objects.klass.name}_grid") do
       context.instance_eval do
